// src/utils/errors/common.d.ts
declare const commonErrorMessages: {
    graphqlError: {
        defaultMessage: string;
        id: string;
    };
    invalid: {
        defaultMessage: string;
        id: string;
    };
    unknownError: {
        defaultMessage: string;
        id: string;
    };
};
export default commonErrorMessages;